refactor custom exceptions into a single module
