+++ pilot/helpers/AgentConvo.py
@@ -5,7 +5,7 @@ import uuid
 from utils.style import color_yellow, color_yellow_bold
 
 from database.database import get_saved_development_step, save_development_step, delete_all_subsequent_steps
-from helpers.exceptions.TokenLimitError import TokenLimitError
+from helpers.exceptions import TokenLimitError
 from utils.function_calling import parse_agent_response, FunctionCallSet
 from utils.llm_connection import create_gpt_chat_completion
 from utils.utils import get_prompt, get_sys_message, capitalize_first_word_with_underscores
